+++ src/get-macros.js
@@ -1,46 +1,42 @@
 "use strict";
-
-
+const vscode = require('vscode')
+/**
+ * get macros in document
+ * @param {vscode.TextDocument} document 
+ * @param {string[]} config 
+ * @returns 
+ */
 function getMacros(document,config){
     let macrosArray = [];
-    let macrosLineIndex = [];
-    if (!document) return {macrosArray, macrosLineIndex}
-
+    if (!document) 
+        return macrosArray
+        
     for (let index = 0; index < document.lineCount; index++) {
         let lineText = document.getText(document.lineAt(index).range)
 
         while (true) {
-            let matchMacro = lineText.match(/(?<init>\\(?:new|renew|provide)command)\s*(?<star>\*?)\s*(?<cmd>(?:\\\w+|\{[\\\w]+\})(?:\s*\[\d+\]\s*){0,2})(\\\S+|\{(?<carry>.*)\})/)
-            if(!matchMacro) break;
-            if(!!matchMacro.groups.carry && matchMacro.groups.carry.indexOf("}")!== -1){
+            let matchMacro = lineText.match(/(?<init>\\(?:new|renew|provide)command)(?<star>\s*\*?\s*)(?<cmd>(?:\\\w+|\{[\\\w]+\})(?:\s*\[\d+\]\s*){0,2})(?<macro>\\\S+|\{(?<carry>.*)\})/)
+            if(!matchMacro) 
+                break;
+            if(matchMacro.groups?.carry?.includes("}")){
                 let carry = matchMacro.groups.carry
-                let init = matchMacro.groups.init+matchMacro.groups.cmd
                 let depth = 1
-                for (let index = 0; index < carry.length; index++) {
-                    const element = carry[index];
-                    depth += (carry[index] === "{")
-                    depth -= (carry[index] === "}")
+                for (let i = 0; i < carry.length; i++) {
+                    depth += (carry[i] === "{");
+                    depth -= (carry[i] === "}");
                     if (!depth) {
-                        matchMacro[0] = init + "{" + carry.substring(0,index)+"}"
+                        matchMacro.groups.macro =  "{" + carry.substring(0,i)+"}"
                         break;
                     }
                 }
             }
-            if(!!matchMacro.groups.star) matchMacro[0]=matchMacro[0].replace(matchMacro.groups.star, "")
-            macrosArray.push(matchMacro[0])
-            lineText = lineText.substring(matchMacro[0].length-1)
-            if (macrosLineIndex.indexOf(index) === -1) macrosLineIndex.push(index)
+            lineText = lineText.substring((matchMacro.groups.init + matchMacro.groups.star + matchMacro.groups.cmd + matchMacro.groups.macro).length - 1) // Fixed a bug that may cause lineText processing error.
+            macrosArray.push(matchMacro.groups.init + matchMacro.groups.cmd + matchMacro.groups.macro)
         }
     }
-
-    macrosArray.push.apply(macrosArray, config || [])
-
-    return {
-        macrosArray,
-        macrosLineIndex
-    }
+    macrosArray.push(...config||[])   
+    return macrosArray;    
 }
-
 
 module.exports = {
     getMacros